+++ mono_repo/test/shared.dart
@@ -177,3 +177,58 @@ void validateOutput(String fileName, String output) {
     });
   }
 }
+
+Future<void> listReadmeSetup() async {
+  await d.dir('pkg1_dir', [
+    d.file('mono_pkg.yaml', ''),
+    d.file('pubspec.yaml', r'''
+name: pkg1
+environment:
+  sdk: '>=2.12.0 <3.0.0'
+
+dependencies:
+  meta: any
+''')
+  ]).create();
+
+  await d.dir('pkg2_dir', [
+    d.file('mono_pkg.yaml', ''),
+    d.file('pubspec.yaml', r'''
+name: pkg2
+publish_to: none
+version: 1.2.3
+environment:
+  sdk: '>=2.12.0 <3.0.0'
+
+dependencies:
+  meta: any
+''')
+  ]).create();
+
+  await d.dir('pkg3_dir', [
+    d.file('mono_pkg.yaml', ''),
+    d.file('pubspec.yaml', r'''
+name: pkg3
+publish_to: https://some.random.domain
+version: 1.2.3
+environment:
+  sdk: '>=2.12.0 <3.0.0'
+
+dependencies:
+  meta: any
+''')
+  ]).create();
+
+  await d.dir('pkg4_dir', [
+    d.file('mono_pkg.yaml', ''),
+    d.file('pubspec.yaml', r'''
+name: pkg4
+version: 1.2.3
+environment:
+  sdk: '>=2.12.0 <3.0.0'
+
+dependencies:
+  meta: any
+''')
+  ]).create();
+}